+++ amplitude.js
@@ -184,7 +184,7 @@ Amplitude.prototype.Identify = Identify;
  *   - includeUtm (boolean) Whether to send utm parameters with events. Defaults to false.
  *   - includeReferrer (boolean) Whether to send referrer info with events. Defaults to false.
  */
-Amplitude.prototype.init = function(apiKey, opt_userId, opt_config) {
+Amplitude.prototype.init = function(apiKey, opt_userId, opt_config, callback) {
   try {
     this.options.apiKey = apiKey;
     if (opt_config) {
@@ -254,6 +254,10 @@ Amplitude.prototype.init = function(apiKey, opt_userId, opt_config) {
     localStorage.setItem(LocalStorageKeys.LAST_EVENT_TIME, this._lastEventTime);
   } catch (e) {
     log(e);
+  }
+
+  if (callback && typeof(callback) === 'function') {
+    callback();
   }
 };
 
@@ -1690,7 +1694,22 @@ module.exports = {
  */
 var localStorage; // jshint ignore:line
 
-if (window.localStorage) {
+// test that Window.localStorage is available and works
+function windowLocalStorageAvailable() {
+  var uid = new Date();
+  var result;
+  try {
+    window.localStorage.setItem(uid, uid);
+    result = window.localStorage.getItem(uid) === String(uid);
+    window.localStorage.removeItem(uid);
+    return result;
+  } catch (e) {
+    // localStorage not available
+  }
+  return false;
+}
+
+if (windowLocalStorageAvailable()) {
   localStorage = window.localStorage;
 } else if (window.globalStorage) {
   // Firefox 2-3 use globalStorage