+++ src/amplitude.js
@@ -72,7 +72,7 @@ Amplitude.prototype.Identify = Identify;
  *   - includeUtm (boolean) Whether to send utm parameters with events. Defaults to false.
  *   - includeReferrer (boolean) Whether to send referrer info with events. Defaults to false.
  */
-Amplitude.prototype.init = function(apiKey, opt_userId, opt_config) {
+Amplitude.prototype.init = function(apiKey, opt_userId, opt_config, callback) {
   try {
     this.options.apiKey = apiKey;
     if (opt_config) {
@@ -142,6 +142,10 @@ Amplitude.prototype.init = function(apiKey, opt_userId, opt_config) {
     localStorage.setItem(LocalStorageKeys.LAST_EVENT_TIME, this._lastEventTime);
   } catch (e) {
     log(e);
+  }
+
+  if (callback && typeof(callback) === 'function') {
+    callback();
   }
 };